Function to ensure compatibility with EpiEstim versions <2.0

Please only use for compatibility; Prefer the new discr_si() function instead

Usage

DiscrSI(k, mu, sigma)

Arguments

k

Positive integer, or vector of positive integers for which the discrete distribution is desired.

mu

A positive real giving the mean of the Gamma distribution.

sigma

A non-negative real giving the standard deviation of the Gamma distribution.