Process incidence input for multivariant analyses

Process incidence input for multivariant analyses with estimate_advantage()

Usage

process_I_multivariant(incid, incid_imported = NULL)

Arguments

incid

a multidimensional array containing values of the incidence for each time step (1st dimension), location (2nd dimension) and pathogen/strain/variant (3rd dimension)

incid_imported

an optional multidimensional array containing values of the incidence of imported cases for each time step (1st dimension), location (2nd dimension) and pathogen/strain/variant (3rd dimension). incid - incid_imported is therefore the incidence of locally infected cases. If incid_imported is NULL this means there are no known imported cases and all cases other than on those from the first time step will be considered locally infected.

Value

a list with two multidimensional elements each with three dimensions: timestep, location and pathogen/strain/variant:

local: an array of the incidence of locally infected cases

imported: an array of the incidence of imported cases

Examples

n_v <- 3 # 3 variants
n_loc <- 1 # 1 location
T <- 100 # 100 time steps
# constant incidence 10 per day everywhere
incid <- array(10, dim = c(T, n_loc, n_v))
process_I_multivariant(incid)
#> $local
#> , , 1
#> 
#>        [,1]
#>   [1,]    0
#>   [2,]   10
#>   [3,]   10
#>   [4,]   10
#>   [5,]   10
#>   [6,]   10
#>   [7,]   10
#>   [8,]   10
#>   [9,]   10
#>  [10,]   10
#>  [11,]   10
#>  [12,]   10
#>  [13,]   10
#>  [14,]   10
#>  [15,]   10
#>  [16,]   10
#>  [17,]   10
#>  [18,]   10
#>  [19,]   10
#>  [20,]   10
#>  [21,]   10
#>  [22,]   10
#>  [23,]   10
#>  [24,]   10
#>  [25,]   10
#>  [26,]   10
#>  [27,]   10
#>  [28,]   10
#>  [29,]   10
#>  [30,]   10
#>  [31,]   10
#>  [32,]   10
#>  [33,]   10
#>  [34,]   10
#>  [35,]   10
#>  [36,]   10
#>  [37,]   10
#>  [38,]   10
#>  [39,]   10
#>  [40,]   10
#>  [41,]   10
#>  [42,]   10
#>  [43,]   10
#>  [44,]   10
#>  [45,]   10
#>  [46,]   10
#>  [47,]   10
#>  [48,]   10
#>  [49,]   10
#>  [50,]   10
#>  [51,]   10
#>  [52,]   10
#>  [53,]   10
#>  [54,]   10
#>  [55,]   10
#>  [56,]   10
#>  [57,]   10
#>  [58,]   10
#>  [59,]   10
#>  [60,]   10
#>  [61,]   10
#>  [62,]   10
#>  [63,]   10
#>  [64,]   10
#>  [65,]   10
#>  [66,]   10
#>  [67,]   10
#>  [68,]   10
#>  [69,]   10
#>  [70,]   10
#>  [71,]   10
#>  [72,]   10
#>  [73,]   10
#>  [74,]   10
#>  [75,]   10
#>  [76,]   10
#>  [77,]   10
#>  [78,]   10
#>  [79,]   10
#>  [80,]   10
#>  [81,]   10
#>  [82,]   10
#>  [83,]   10
#>  [84,]   10
#>  [85,]   10
#>  [86,]   10
#>  [87,]   10
#>  [88,]   10
#>  [89,]   10
#>  [90,]   10
#>  [91,]   10
#>  [92,]   10
#>  [93,]   10
#>  [94,]   10
#>  [95,]   10
#>  [96,]   10
#>  [97,]   10
#>  [98,]   10
#>  [99,]   10
#> [100,]   10
#> 
#> , , 2
#> 
#>        [,1]
#>   [1,]    0
#>   [2,]   10
#>   [3,]   10
#>   [4,]   10
#>   [5,]   10
#>   [6,]   10
#>   [7,]   10
#>   [8,]   10
#>   [9,]   10
#>  [10,]   10
#>  [11,]   10
#>  [12,]   10
#>  [13,]   10
#>  [14,]   10
#>  [15,]   10
#>  [16,]   10
#>  [17,]   10
#>  [18,]   10
#>  [19,]   10
#>  [20,]   10
#>  [21,]   10
#>  [22,]   10
#>  [23,]   10
#>  [24,]   10
#>  [25,]   10
#>  [26,]   10
#>  [27,]   10
#>  [28,]   10
#>  [29,]   10
#>  [30,]   10
#>  [31,]   10
#>  [32,]   10
#>  [33,]   10
#>  [34,]   10
#>  [35,]   10
#>  [36,]   10
#>  [37,]   10
#>  [38,]   10
#>  [39,]   10
#>  [40,]   10
#>  [41,]   10
#>  [42,]   10
#>  [43,]   10
#>  [44,]   10
#>  [45,]   10
#>  [46,]   10
#>  [47,]   10
#>  [48,]   10
#>  [49,]   10
#>  [50,]   10
#>  [51,]   10
#>  [52,]   10
#>  [53,]   10
#>  [54,]   10
#>  [55,]   10
#>  [56,]   10
#>  [57,]   10
#>  [58,]   10
#>  [59,]   10
#>  [60,]   10
#>  [61,]   10
#>  [62,]   10
#>  [63,]   10
#>  [64,]   10
#>  [65,]   10
#>  [66,]   10
#>  [67,]   10
#>  [68,]   10
#>  [69,]   10
#>  [70,]   10
#>  [71,]   10
#>  [72,]   10
#>  [73,]   10
#>  [74,]   10
#>  [75,]   10
#>  [76,]   10
#>  [77,]   10
#>  [78,]   10
#>  [79,]   10
#>  [80,]   10
#>  [81,]   10
#>  [82,]   10
#>  [83,]   10
#>  [84,]   10
#>  [85,]   10
#>  [86,]   10
#>  [87,]   10
#>  [88,]   10
#>  [89,]   10
#>  [90,]   10
#>  [91,]   10
#>  [92,]   10
#>  [93,]   10
#>  [94,]   10
#>  [95,]   10
#>  [96,]   10
#>  [97,]   10
#>  [98,]   10
#>  [99,]   10
#> [100,]   10
#> 
#> , , 3
#> 
#>        [,1]
#>   [1,]    0
#>   [2,]   10
#>   [3,]   10
#>   [4,]   10
#>   [5,]   10
#>   [6,]   10
#>   [7,]   10
#>   [8,]   10
#>   [9,]   10
#>  [10,]   10
#>  [11,]   10
#>  [12,]   10
#>  [13,]   10
#>  [14,]   10
#>  [15,]   10
#>  [16,]   10
#>  [17,]   10
#>  [18,]   10
#>  [19,]   10
#>  [20,]   10
#>  [21,]   10
#>  [22,]   10
#>  [23,]   10
#>  [24,]   10
#>  [25,]   10
#>  [26,]   10
#>  [27,]   10
#>  [28,]   10
#>  [29,]   10
#>  [30,]   10
#>  [31,]   10
#>  [32,]   10
#>  [33,]   10
#>  [34,]   10
#>  [35,]   10
#>  [36,]   10
#>  [37,]   10
#>  [38,]   10
#>  [39,]   10
#>  [40,]   10
#>  [41,]   10
#>  [42,]   10
#>  [43,]   10
#>  [44,]   10
#>  [45,]   10
#>  [46,]   10
#>  [47,]   10
#>  [48,]   10
#>  [49,]   10
#>  [50,]   10
#>  [51,]   10
#>  [52,]   10
#>  [53,]   10
#>  [54,]   10
#>  [55,]   10
#>  [56,]   10
#>  [57,]   10
#>  [58,]   10
#>  [59,]   10
#>  [60,]   10
#>  [61,]   10
#>  [62,]   10
#>  [63,]   10
#>  [64,]   10
#>  [65,]   10
#>  [66,]   10
#>  [67,]   10
#>  [68,]   10
#>  [69,]   10
#>  [70,]   10
#>  [71,]   10
#>  [72,]   10
#>  [73,]   10
#>  [74,]   10
#>  [75,]   10
#>  [76,]   10
#>  [77,]   10
#>  [78,]   10
#>  [79,]   10
#>  [80,]   10
#>  [81,]   10
#>  [82,]   10
#>  [83,]   10
#>  [84,]   10
#>  [85,]   10
#>  [86,]   10
#>  [87,]   10
#>  [88,]   10
#>  [89,]   10
#>  [90,]   10
#>  [91,]   10
#>  [92,]   10
#>  [93,]   10
#>  [94,]   10
#>  [95,]   10
#>  [96,]   10
#>  [97,]   10
#>  [98,]   10
#>  [99,]   10
#> [100,]   10
#> 
#> 
#> $imported
#> , , 1
#> 
#>        [,1]
#>   [1,]   10
#>   [2,]    0
#>   [3,]    0
#>   [4,]    0
#>   [5,]    0
#>   [6,]    0
#>   [7,]    0
#>   [8,]    0
#>   [9,]    0
#>  [10,]    0
#>  [11,]    0
#>  [12,]    0
#>  [13,]    0
#>  [14,]    0
#>  [15,]    0
#>  [16,]    0
#>  [17,]    0
#>  [18,]    0
#>  [19,]    0
#>  [20,]    0
#>  [21,]    0
#>  [22,]    0
#>  [23,]    0
#>  [24,]    0
#>  [25,]    0
#>  [26,]    0
#>  [27,]    0
#>  [28,]    0
#>  [29,]    0
#>  [30,]    0
#>  [31,]    0
#>  [32,]    0
#>  [33,]    0
#>  [34,]    0
#>  [35,]    0
#>  [36,]    0
#>  [37,]    0
#>  [38,]    0
#>  [39,]    0
#>  [40,]    0
#>  [41,]    0
#>  [42,]    0
#>  [43,]    0
#>  [44,]    0
#>  [45,]    0
#>  [46,]    0
#>  [47,]    0
#>  [48,]    0
#>  [49,]    0
#>  [50,]    0
#>  [51,]    0
#>  [52,]    0
#>  [53,]    0
#>  [54,]    0
#>  [55,]    0
#>  [56,]    0
#>  [57,]    0
#>  [58,]    0
#>  [59,]    0
#>  [60,]    0
#>  [61,]    0
#>  [62,]    0
#>  [63,]    0
#>  [64,]    0
#>  [65,]    0
#>  [66,]    0
#>  [67,]    0
#>  [68,]    0
#>  [69,]    0
#>  [70,]    0
#>  [71,]    0
#>  [72,]    0
#>  [73,]    0
#>  [74,]    0
#>  [75,]    0
#>  [76,]    0
#>  [77,]    0
#>  [78,]    0
#>  [79,]    0
#>  [80,]    0
#>  [81,]    0
#>  [82,]    0
#>  [83,]    0
#>  [84,]    0
#>  [85,]    0
#>  [86,]    0
#>  [87,]    0
#>  [88,]    0
#>  [89,]    0
#>  [90,]    0
#>  [91,]    0
#>  [92,]    0
#>  [93,]    0
#>  [94,]    0
#>  [95,]    0
#>  [96,]    0
#>  [97,]    0
#>  [98,]    0
#>  [99,]    0
#> [100,]    0
#> 
#> , , 2
#> 
#>        [,1]
#>   [1,]   10
#>   [2,]    0
#>   [3,]    0
#>   [4,]    0
#>   [5,]    0
#>   [6,]    0
#>   [7,]    0
#>   [8,]    0
#>   [9,]    0
#>  [10,]    0
#>  [11,]    0
#>  [12,]    0
#>  [13,]    0
#>  [14,]    0
#>  [15,]    0
#>  [16,]    0
#>  [17,]    0
#>  [18,]    0
#>  [19,]    0
#>  [20,]    0
#>  [21,]    0
#>  [22,]    0
#>  [23,]    0
#>  [24,]    0
#>  [25,]    0
#>  [26,]    0
#>  [27,]    0
#>  [28,]    0
#>  [29,]    0
#>  [30,]    0
#>  [31,]    0
#>  [32,]    0
#>  [33,]    0
#>  [34,]    0
#>  [35,]    0
#>  [36,]    0
#>  [37,]    0
#>  [38,]    0
#>  [39,]    0
#>  [40,]    0
#>  [41,]    0
#>  [42,]    0
#>  [43,]    0
#>  [44,]    0
#>  [45,]    0
#>  [46,]    0
#>  [47,]    0
#>  [48,]    0
#>  [49,]    0
#>  [50,]    0
#>  [51,]    0
#>  [52,]    0
#>  [53,]    0
#>  [54,]    0
#>  [55,]    0
#>  [56,]    0
#>  [57,]    0
#>  [58,]    0
#>  [59,]    0
#>  [60,]    0
#>  [61,]    0
#>  [62,]    0
#>  [63,]    0
#>  [64,]    0
#>  [65,]    0
#>  [66,]    0
#>  [67,]    0
#>  [68,]    0
#>  [69,]    0
#>  [70,]    0
#>  [71,]    0
#>  [72,]    0
#>  [73,]    0
#>  [74,]    0
#>  [75,]    0
#>  [76,]    0
#>  [77,]    0
#>  [78,]    0
#>  [79,]    0
#>  [80,]    0
#>  [81,]    0
#>  [82,]    0
#>  [83,]    0
#>  [84,]    0
#>  [85,]    0
#>  [86,]    0
#>  [87,]    0
#>  [88,]    0
#>  [89,]    0
#>  [90,]    0
#>  [91,]    0
#>  [92,]    0
#>  [93,]    0
#>  [94,]    0
#>  [95,]    0
#>  [96,]    0
#>  [97,]    0
#>  [98,]    0
#>  [99,]    0
#> [100,]    0
#> 
#> , , 3
#> 
#>        [,1]
#>   [1,]   10
#>   [2,]    0
#>   [3,]    0
#>   [4,]    0
#>   [5,]    0
#>   [6,]    0
#>   [7,]    0
#>   [8,]    0
#>   [9,]    0
#>  [10,]    0
#>  [11,]    0
#>  [12,]    0
#>  [13,]    0
#>  [14,]    0
#>  [15,]    0
#>  [16,]    0
#>  [17,]    0
#>  [18,]    0
#>  [19,]    0
#>  [20,]    0
#>  [21,]    0
#>  [22,]    0
#>  [23,]    0
#>  [24,]    0
#>  [25,]    0
#>  [26,]    0
#>  [27,]    0
#>  [28,]    0
#>  [29,]    0
#>  [30,]    0
#>  [31,]    0
#>  [32,]    0
#>  [33,]    0
#>  [34,]    0
#>  [35,]    0
#>  [36,]    0
#>  [37,]    0
#>  [38,]    0
#>  [39,]    0
#>  [40,]    0
#>  [41,]    0
#>  [42,]    0
#>  [43,]    0
#>  [44,]    0
#>  [45,]    0
#>  [46,]    0
#>  [47,]    0
#>  [48,]    0
#>  [49,]    0
#>  [50,]    0
#>  [51,]    0
#>  [52,]    0
#>  [53,]    0
#>  [54,]    0
#>  [55,]    0
#>  [56,]    0
#>  [57,]    0
#>  [58,]    0
#>  [59,]    0
#>  [60,]    0
#>  [61,]    0
#>  [62,]    0
#>  [63,]    0
#>  [64,]    0
#>  [65,]    0
#>  [66,]    0
#>  [67,]    0
#>  [68,]    0
#>  [69,]    0
#>  [70,]    0
#>  [71,]    0
#>  [72,]    0
#>  [73,]    0
#>  [74,]    0
#>  [75,]    0
#>  [76,]    0
#>  [77,]    0
#>  [78,]    0
#>  [79,]    0
#>  [80,]    0
#>  [81,]    0
#>  [82,]    0
#>  [83,]    0
#>  [84,]    0
#>  [85,]    0
#>  [86,]    0
#>  [87,]    0
#>  [88,]    0
#>  [89,]    0
#>  [90,]    0
#>  [91,]    0
#>  [92,]    0
#>  [93,]    0
#>  [94,]    0
#>  [95,]    0
#>  [96,]    0
#>  [97,]    0
#>  [98,]    0
#>  [99,]    0
#> [100,]    0
#> 
#> 
#> attr(,"class")
#> [1] "incid_multivariant"