EpiEstim: Estimate Time Varying Reproduction Numbers from Epidemic Curves

The EpiEstim package provides tools to Tools to quantify transmissibility throughout an epidemic from the analysis of time series of incidence.

Bibliography

A BibTeX file of papers that describe the methodology underlying EpiEstim is available via system.file("epiestimpapers.bib", package = "EpiEstim").

See also

Author

Maintainer: Anne Cori a.cori@imperial.ac.uk (ORCID)

Other contributors:

Simon Cauchemez [contributor]

Neil M. Ferguson (ORCID) [contributor]

Christophe Fraser (ORCID) [contributor]

Elisabeth Dahlqwist (ORCID) [contributor]

P. Alex Demarsh [contributor]

Thibaut Jombart (ORCID) [contributor]

Zhian N. Kamvar (ORCID) [contributor]

Justin Lessler (ORCID) [contributor]

Shikun Li [contributor]

Jonathan A. Polonsky (ORCID) [contributor]

Jake Stockwin [contributor]

Robin Thompson (ORCID) [contributor]

Rolina van Gaalen [contributor]

Rebecca Nash r.nash@imperial.ac.uk (ORCID) [contributor]

Sangeeta Bhatia (ORCID) [contributor]

Jack Wardle (ORCID) [contributor]

Andrea Brizzi (ORCID) [contributor]