Compute the smallest index at which joint estimation should start

Unless specified by the user, t_min in estimate_advantage is computed as the sum of two indices: (i) the first day of non-zero incidence across all locations, computed using first_nonzero_incid and (ii) the 95th percentile of the probability mass function of the SI distribution across all variants computed using compute_si_cutoff

compute_t_min(incid, si_distr, miss_at_most)

Arguments

incid
a multidimensional array containing values of the incidence for each time step (1st dimension), location (2nd dimension) and pathogen/strain/variant (3rd dimension)

si_distr
a matrix with two columns, each containing the probability mass function for the discrete serial interval for each of the two pathogen/strain/variants, starting with the probability mass function for day 0 in the first row, which should be 0. each column in the matrix should sum to 1

miss_at_most
numeric. probability mass in the tail of the SI distribution

Value

integer

Author

Sangeeta Bhatia